+++ .pylintrc
@@ -1,5 +1,5 @@
 [BASIC]
-good-names=i, j, k, ir, e
+good-names=i, j, k, ir, e, tb
 
 [FORMAT]
 # variable-rgx="[a-z_][a-z0-9_]{0,30}$"
