+++ python2verilog/api/__init__.py
@@ -4,8 +4,8 @@ API wrappers
 
 
 from . import exit_handler  # otherwise file is not loaded
+from .from_context import context_to_text
+from .from_text import text_to_text
 from .modes import Modes
 from .namespace import new_namespace
-from .text import text_to_text
 from .verilogify import verilogify
-from .wrappers import context_to_text
